Extract VM from cell details. add new sales channel label to summary
